Post관련 READ api 작성
pagenation 해서, 글 목록 넘기는 API와
글 하나의 세부 내용을 넘기는 API 작성

+++ controllers/postController.js
@@ -6,8 +6,12 @@ const writePost = async (req, res, next) => {
         const post_info = req.body; // title, desc
         post_info.createAt = Date.now();
         post_info.user_id = req.user;
-        await PostModel.createPost(post_info);
-        res.status(201).json({result : true});
+        const post_id = await PostModel.createPost(post_info);
+        res.status(201).json({
+            result : true,
+            post_id,
+            message: '글이 등록되었습니다.'
+        });
     } catch(err){
         next(err);
     }
@@ -15,7 +19,7 @@ const writePost = async (req, res, next) => {
 const modifyPost = async (req, res, next) => {
     try {
         if (req.user == null) throw 'Unauthorized';
-        const post_info = req.body; // title, desc
+        const post_info = req.body;
         const user_id = req.user;
         post_info.updateAt = Date.now();
         await PostModel.updatePost(post_info, user_id);
@@ -24,14 +28,34 @@ const modifyPost = async (req, res, next) => {
         next(err);
     }
 }
-const deletePost = async (req,res,next) => {
+const deletePost = async (req, res, next) => {
     try {
-        const post_id = req.params.postId;
-        console.log(post_id);
+        const post_id = req.params.post_id;
         const user_id = req.user;
-        console.log(user_id);
+        if (req.user == null) throw 'Unauthorized';
         await PostModel.deletePost(post_id, user_id);
         res.status(204).send();
+    } catch(err){
+        next(err);
+    }
+}
+const getPostList = async (req, res, next) => {
+    try {
+        const page_num = req.params.page_num;
+        const post_num = 10; // 한 번에 보여줄 글 개수
+        const post_list = await PostModel.selectPosts(page_num, post_num);
+        const simple_list = post_list.map((post) => {return {_id: post._id, title: post.title}});
+        res.status(200).json(simple_list);
+    } catch(err){
+        next(err);
+    }
+}
+const getPostDetail = async (req, res, next) => {
+    try {
+        const post_id = req.params.post_id;
+        const post = await PostModel.selectPost(post_id);
+        res.status(200).json(post);
+        
     } catch(err){
         next(err);
     }
@@ -40,5 +64,6 @@ module.exports = {
     writePost,
     modifyPost,
     deletePost,
-
+    getPostList,
+    getPostDetail
 }
